+++ vst/protocol/protocol_version.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2017 ArangoDB GmbH, Cologne, Germany
+// Copyright 2017-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,10 +17,12 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Ewout Prangsma
-//
 
 package protocol
+
+import (
+	"fmt"
+)
 
 // Version indicates the version of the Velocystream protocol
 type Version int
@@ -29,3 +31,14 @@ const (
 	Version1_0 Version = iota // VST 1.0
 	Version1_1                // VST 1.1
 )
+
+func (v Version) asString() (string, error) {
+	switch v {
+	case Version1_0:
+		return "vst/1.0", nil
+	case Version1_1:
+		return "vst/1.1", nil
+	default:
+		return "", fmt.Errorf("unknown protocol version %d", int(v))
+	}
+}